+++ src/Controller/PaymentController.php
@@ -134,6 +134,7 @@ class PaymentController extends PaymentController_parent
          * check:
          * - all none PP-Payments
          * - payPalHealth
+         * - deprecated payment method
          * - currency
          * - country
          * - netto-mode
@@ -144,6 +145,8 @@ class PaymentController extends PaymentController_parent
                 !isset($payPalDefinitions[$key]) ||
                 (
                     $payPalHealth &&
+                    //dont add payment that is deprecated
+                    ( !PayPalDefinitions::isDeprecatedPayment($payment->getId()) ) &&
                     (
                         empty($payPalDefinitions[$key]['currencies']) ||
                         in_array($actShopCurrency->name, $payPalDefinitions[$key]['currencies'], true)