+++ src/Core/PayPalDefinitions.php
@@ -380,7 +380,8 @@ final class PayPalDefinitions
             'constraints' => self::PAYMENT_CONSTRAINTS_UAPM,
             'onlybrutto' => false,
             'buttonpayment' => false,
-            'defaulton' => true
+            'defaulton' => false,
+            'deprecated' => true,
         ],
         // uAPM iDEAL
         'oscpaypal_ideal' => [
@@ -512,5 +513,19 @@ final class PayPalDefinitions
             isset(self::PAYPAL_DEFINTIONS[$paymentId]['vaultingtype'])
             && self::PAYPAL_DEFINTIONS[$paymentId]['vaultingtype'] === $paypalPaymentType
         );
+    }
+
+    /**
+     * Check if payment is deprecated
+     *
+     * @param string $paymentId
+     * @return bool
+     */
+    public static function isDeprecatedPayment(string $paymentId): bool
+    {
+        if ( isset(self::PAYPAL_DEFINTIONS[$paymentId]['deprecated']) && self::PAYPAL_DEFINTIONS[$paymentId]['deprecated'] === true ) {
+            return true;
+        }
+        return false;
     }
 }